Add name to childApp instance and getName method
With this, now childApp instances will have their name attached to instance itself. A appName can be retrieved by using the new getName method. If an app is a childApp then the appName will be returned. However if the app is not a childApp or is a parentApp then undefined will be returned. This commit includes src code, tests and docs.