Move view, region, and state events to running only events
This is breaking but much more reasonable.

The current setup allows for too much unexpected behavior when an app is not running.  Essentially setting up apps can be done with the app is not running, but the app API should cater to running apps almost across the board.

This makes region/view behave similarly to the state model for the app.