Introduce Next Component
This component keeps the idea but greatly reduces the API, removing awkward issues where `renderView` could be called previous to `show`  This also allows for a component to construct and reconstruct it's view without destroying the component.  The component also is now built from `Marionette.Application` which allows it to utilize the existing region/view API, so `getRegion` needs to be used over `.region` and `getView()` must be used over `.currentView` which was removed.

This is a breaking change but it should be a minor one.